requirements.txt
Django==4.2.1
python-dotenv==0.21.1
gunicorn==20.1.0
djangorestframework==3.14.0
